spread the thermo and transport into separate files
