+++ reactorch/solution.py
@@ -19,12 +19,27 @@ torch.set_default_tensor_type("torch.DoubleTensor")
 class Solution(nn.Module):
     from .import_kinetics import set_nasa
     from .import_kinetics import set_reactions
+    from .import_kinetics import set_transport
 
     from .kinetics import forward_rate_constants_func
     from .kinetics import forward_rate_constants_func_vec
     from .kinetics import equilibrium_constants_func
     from .kinetics import reverse_rate_constants_func
     from .kinetics import wdot_func
+
+    from .thermo import cp_mole_func
+    from .thermo import cp_mass_func
+    from .thermo import enthalpy_mole_func
+    from .thermo import enthalpy_mass_func
+    from .thermo import entropy_mole_func
+    from .thermo import entropy_mass_func
+
+    from .transport import update_transport
+    from .transport import viscosities_func
+    from .transport import thermal_conductivity_func
+    from .transport import binary_diff_coeffs_func
+
+    from .magic_function import C2X, Y2X, Y2C, X2C, X2Y
 
     def __init__(self, mech_yaml=None, device=None, vectorize=False):
         super(Solution, self).__init__()
@@ -41,7 +56,7 @@ class Solution(nn.Module):
 
         self.R = ct.gas_constant
 
-        self.P_atm = torch.Tensor([ct.one_atm]).to(self.device)
+        self.one_atm = torch.Tensor([ct.one_atm]).to(self.device)
 
         self.n_species = self.gas.n_species
 
@@ -120,144 +135,6 @@ class Solution(nn.Module):
         self.reverse_rate_constants_func()
 
         self.wdot_func()
-
-    def update_transport(self):
-
-        self.viscosities_func()
-
-        self.thermal_conductivity_func()
-
-        self.binary_diff_coeffs_func()
-
-    def set_transport(self, species_viscosities_poly, thermal_conductivity_poly,
-                      binary_diff_coeffs_poly):
-        # Transport Properties
-
-        self.species_viscosities_poly = torch.from_numpy(species_viscosities_poly).to(self.device)
-
-        self.thermal_conductivity_poly = torch.from_numpy(thermal_conductivity_poly).to(self.device)
-
-        self.binary_diff_coeffs_poly = torch.from_numpy(binary_diff_coeffs_poly).to(self.device)
-
-        self.poly_order = self.species_viscosities_poly.shape[0]
-
-    def viscosities_func(self):
-
-        self.trans_T = torch.cat([self.logT ** i for i in reversed(range(self.poly_order))], dim=1)
-
-        self.species_viscosities = torch.mm(self.trans_T, self.species_viscosities_poly)
-
-        self.Wk_over_Wj = torch.mm(self.molecular_weights, 1 / self.molecular_weights.T)
-
-        self.Wj_over_Wk = 1 / self.Wk_over_Wj
-
-        self.etak_over_etaj = torch.bmm(
-            self.species_viscosities.unsqueeze(-1),
-            1 / self.species_viscosities.unsqueeze(-1).view(-1, 1, self.n_species))
-
-        self.PHI = (1 / 2.8284271247461903 / torch.sqrt(1 + self.Wk_over_Wj) *
-                    (1 + self.etak_over_etaj ** 0.5 * self.Wj_over_Wk ** 0.25) ** 2)
-
-        self.viscosities = (self.X.clone() * self.species_viscosities /
-                            torch.bmm(self.PHI, self.X.unsqueeze(-1)).squeeze(-1)).sum(dim=1,
-                                                                                       keepdim=True)
-
-    def thermal_conductivity_func(self):
-
-        self.species_thermal_conductivity = torch.mm(self.trans_T, self.thermal_conductivity_poly)
-
-        self.thermal_conductivity = 0.5 * (
-            (self.X.clone() * self.species_thermal_conductivity).sum(dim=1, keepdim=True) +
-            1 / (self.X.clone() / self.species_thermal_conductivity).sum(dim=1, keepdim=True)
-        )
-
-    def binary_diff_coeffs_func(self):
-
-        self.binary_diff_coeffs = torch.mm(
-            self.trans_T, self.binary_diff_coeffs_poly).view(-1, self.n_species, self.n_species)
-
-        self.X_eps = self.X.clamp_(1e-12)
-
-        self.XjWj = (torch.mm(self.X_eps, self.molecular_weights) -
-                     self.X_eps * self.molecular_weights.T)
-
-        self.XjDjk = torch.bmm(self.X_eps.view(-1, 1, self.n_species), 1 / self.binary_diff_coeffs).squeeze(
-            1) - self.X_eps / self.binary_diff_coeffs.diagonal(dim1=-2, dim2=-1)
-
-        self.mix_diff_coeffs = self.XjWj / self.XjDjk / self.mean_molecular_weight / self.P * self.P_atm
-
-    # Magic Functions
-
-    def C2X(self):
-        self.X = (self.C.T / self.C.sum(dim=1)).T
-
-    def Y2X(self):
-        self.X = self.Y * self.mean_molecular_weight / self.molecular_weights.T
-
-    def Y2C(self):
-        self.C = self.Y * self.density_mass / self.molecular_weights.T
-
-    def X2C(self):
-        self.C = self.X * self.density_mass / self.mean_molecular_weight
-
-    def X2Y(self):
-        self.Y = self.X * self.molecular_weights.T / self.mean_molecular_weight
-
-    def cp_mole_func(self):
-
-        self.cp_T = torch.cat([self.T ** 0, self.T, self.T ** 2, self.T ** 3, self.T ** 4], dim=1)
-
-        self.cp = (torch.mm(self.cp_T, self.nasa_low[:, :5].T) * (self.T <= 1000).int() +
-                   torch.mm(self.cp_T, self.nasa_high[:, :5].T) * (self.T > 1000).int())
-
-        self.cp_mole = (self.R * self.cp * self.X.clone()).sum(dim=1)
-
-    def cp_mass_func(self):
-
-        self.cp_mass = self.cp_mole / self.mean_molecular_weight.T
-
-    def enthalpy_mole_func(self):
-
-        self.H_T = torch.cat((self.T ** 0, self.T / 2, self.T ** 2 / 3,
-                              self.T ** 3 / 4, self.T ** 4 / 5, 1 / self.T), dim=1)
-
-        self.H = (torch.mm(self.H_T, self.nasa_low[:, :6].T) * (self.T <= 1000).int() +
-                  torch.mm(self.H_T, self.nasa_high[:, :6].T) * (self.T > 1000).int())
-
-        self.H = self.H * self.R * self.T
-
-        self.partial_molar_enthalpies = self.H
-
-        self.enthalpy_mole = (self.H * self.X).sum(dim=1)
-
-    def enthalpy_mass_func(self):
-
-        self.h = self.H / self.molecular_weights.T
-
-        self.enthalpy_mass = (self.h * self.Y).sum(dim=1)
-
-    def entropy_mole_func(self):
-
-        self.S_T = torch.cat((torch.log(self.T), self.T, self.T ** 2 / 2,
-                              self.T ** 3 / 3, self.T ** 4 / 4, self.T ** 0), dim=1)
-
-        self.S0 = (
-            torch.mm(self.S_T, self.nasa_low[:, [0, 1, 2, 3, 4, 6]].T) * (self.T <= 1000).int() +
-            torch.mm(self.S_T, self.nasa_high[:, [0, 1, 2, 3, 4, 6]].T) * (self.T > 1000).int())
-
-        self.S0 = self.S0 * self.R
-
-        self.S = self.S0 - self.R * torch.log(self.X) - self.R * torch.log(self.P / ct.one_atm)
-
-        self.partial_molar_entropies = self.S
-
-        self.entropy_mole = (self.S * self.X).sum(dim=1)
-
-    def entropy_mass_func(self):
-
-        self.s = self.S / self.molecular_weights.T
-
-        self.entropy_mass = (self.s * self.Y).sum(dim=1)
 
     def Ydot_func(self):
